### 2019-09-01 #1697 @whedon
[PRE REVIEW]: ndsplines: A Python Library for Tensor-Product B-Splines of Arbitrary Dimension
**Submitting author:** @sixpearls (<a href="http://orcid.org/0000-0001-5602-1888">Benjamin Margolis</a>)
**Repository:** <a href="https://github.com/kb-press/ndsplines/" target ="_blank">https://github.com/kb-press/ndsplines/</a>
**Version:** v0.1.0
