### 2019-09-01 #1697 @labarba
👋  @mbobra — Could you handle this JOSS submission as editor?
"ndsplines: A Python Library for Tensor-Product B-Splines of Arbitrary Dimension"
repo: https://github.com/kb-press/ndsplines/
