### 2019-09-01 #1697 @sixpearls
@labarba Thanks, I will correct the bibliography.
