### 2019-09-01 e9ca2d7
paper corrections
