+++ paper/paper.md
@@ -22,7 +22,7 @@ bibliography: references.bib
 
 In order to use data driven models in analysis, the data must be projected onto a set of basis functions to allow the model to be evaluated over a continuous domain. Additionally, the known form of the basis functions enables the use of calculus and other symbolic operations to be performed on the representation of the data model.
 
-B-Splines are a particularly useful basis because they have beneficial analytic properties and because many efficient algorithms for their creation, manipulation, and evaluation have already been developed [@deboor]. Some of the beneficial properties of the B-Spline basis functions include: a closed support, smoothness, and amenability to efficient and stable algorithms. Strictly speaking, the B-Spline basis functions are 1-dimensional maps from the reals to the reals. However, they can be readily extended to N-dimensions by taking tensor products of B-Splines on orthogonal 1-dimensional spaces for each variable. 
+B-Splines are a particularly useful basis because they have beneficial analytic properties and because many efficient algorithms for their creation, manipulation, and evaluation have already been developed [@de1978practical]. Some of the beneficial properties of the B-Spline basis functions include: a closed support, smoothness, and amenability to efficient and stable algorithms. Strictly speaking, the B-Spline basis functions are 1-dimensional maps from the reals to the reals. However, they can be readily extended to N-dimensions by taking tensor products of B-Splines on orthogonal 1-dimensional spaces for each variable. 
 
 B-Splines are well studied in the field of computer graphics where they are particularly used in computer aided-design (CAD) and graphic design programs. Due to this popularity, many programming language ecosystems have implementations of 1- and 2-dimensional splines such as might be used to design curves or surfaces. For example, SciPy provides a number of interfaces to Dierckx's FORTRAN implementation as well as a custom 1-D implementation [@scipy, @dierckx1995curve]. However, the authors are only aware of one general N-dimensional splines implementation [@de1990spline].
 
