+++ paper/references.bib
@@ -5,7 +5,8 @@
   volume={2},
   number={17},
   pages={396},
-  year={2017}
+  year={2017},
+  doi= {10.21105/joss.00396},
 }
 
 @Misc{scipy,
@@ -24,7 +25,8 @@
   number={2},
   pages={22},
   year={2011},
-  publisher={IEEE Computer Society}
+  publisher={IEEE Computer Society},
+  doi={10.1109/mcse.2011.37},
 }
 
 @article{behnel2011cython,
@@ -35,7 +37,8 @@
   number={2},
   pages={31},
   year={2011},
-  publisher={IEEE Computer Society}
+  publisher={IEEE Computer Society},
+  doi={10.1109/mcse.2010.118}
 }
 
 @book{dierckx1995curve,
@@ -65,7 +68,8 @@
   author={D'Souza, Sarah N and Okolo, Wendy and Nikaido, Ben and Yount, Bryan and Tran, Jason and Margolis, Benjamin and Smith, Brandon and Cassell, Alan and Johnson, Breanna and Hibbard, Kenneth and others},
   booktitle={AIAA Aviation 2019 Forum},
   pages={2901},
-  year={2019}
+  year={2019},
+  doi={10.2514/6.2019-2901}
 }
 
 @inproceedings{margolis2019control,
@@ -80,5 +84,6 @@
   author={Jackson, E Bruce and Hildreth, Bruce},
   booktitle={AIAA Modeling and Simulation Technologies Conference and Exhibit},
   pages={4482},
-  year={2002}
+  year={2002},
+  doi={10.2514/6.2002-4482}
 }
