+++ paper/references.bib
@@ -57,10 +57,10 @@
 
 @book{de1978practical,
   title={A practical guide to splines},
-  author={De Boor, Carl and De Boor, Carl and Math{\'e}maticien, Etats-Unis and De Boor, Carl and De Boor, Carl},
+  author={De Boor, Carl},
   volume={27},
   year={1978},
-  publisher={springer-verlag New York}
+  publisher={Springer-Verlag New York}
 }
 
 @inproceedings{d2019developing,
