+++ paper/references.bib
@@ -14,7 +14,7 @@
   title =     {{SciPy}: Open source scientific tools for {Python}},
   year =      {2001--},
   url = "http://www.scipy.org/",
-  note = {[Online; accessed <today>]}
+  note = {[Online; accessed 9/1/2019]}
 }
 
 @article{van2011numpy,
@@ -60,7 +60,8 @@
   author={De Boor, Carl},
   volume={27},
   year={1978},
-  publisher={Springer-Verlag New York}
+  publisher={Springer-Verlag New York},
+  doi={10.1002/zamm.19800600129}
 }
 
 @inproceedings{d2019developing,
