### 2019-09-03 #1697 @arfon
I'm going to go ahead and assign @mbobra as the handling editor for this submission. @sixpearls - it's going to be another week or so before @mbobra is back to JOSS editorial activities - your patience is much appreciated :-)
