### 2019-09-16 #1697 @kthyng
Hi @sixpearls did you look through the list of volunteer reviewers to see if you could find any suggestions? [Here](https://bit.ly/joss-reviewers) is the list. Suggesting a handful is most helpful if possible since we need two and people aren't always available.
