### 2019-09-16 #1697 @sixpearls
Hi @vissarion, @amritagos, @hugoledoux, @HaoZeke, @trp07, @sarats, would two of you be willing to act as reviewers for this JOSS submission on N-dimensional splines? The repository is https://github.com/kb-press/ndsplines/, if you would like to check it out
