### 2019-09-18 kb-press/ndsplines#53 @amritagos
[JOSS] License Formatting
Part of https://github.com/openjournals/joss-reviews/issues/1745. Could you make the LICENSE file contain just the BSD-3 clause text for this project? It would be better to mention the other license (SciPy) in the project `README.md`, under an External libraries heading or under a Licenses heading. This is essentially to ensure that the license is instantly visible to whoever visits the repo.

![image](https://user-images.githubusercontent.com/32496815/65140145-8cab5b80-da2b-11e9-9178-d3967bdffb17.png)
