### 2019-10-06 #1745 @amritagos
Hi @sixpearls, I have completed my review and am waiting for the issue to be resolved for my closing comments. Great work!
