### 2019-10-07 kb-press/ndsplines#53 @sixpearls
Hi @amritagos, we have reviewed this request and the options available to have GitHub parse the license file. We believe that the best place for the SciPy copyright/license information is in a dedicated and obvious license file in the root directory. It turns out that even adding a supplemental license file would cause GitHub to use the "See License" button rather than parsing the primary LICENSE file and putting the name in the header.

However, we do agree that it would be convenient to make the license type easier to determine. To address this, we added the name of the license to the top of the LICENSE file. Do you agree that his makes it sufficiently easy to determine the license for anyone who would like to know?
