### 2019-10-07 kb-press/ndsplines#53 @amritagos
Sure, that'll do. Thanks for detailing your fix.
