### 2019-10-07 #1745 @amritagos
@mbobra, I have completed my review and am certain that the work presented by @sixpearls and @ixjlyons is well within the JOSS purview. The work is in accordance with the rules and standards of JOSS and I am glad to recommend this for acceptance and subsequent publication. Thanks!
