### 2019-10-12 241b8c1
Explicitly include pyx files in sdist
