+++ MANIFEST.in
@@ -1,3 +1,4 @@
 include readme.rst
 include LICENSE
 include ndsplines/*.h
+include ndsplines/*.pyx
