### 2019-10-15 #1745 @mbobra
@sixpearls We're almost there! Can you please [archive your release on Zenodo](https://guides.github.com/activities/citable-code/) to obtain a DOI and then put that DOI in your `readme.rst` file? Please make sure that the title and author list on the Zenodo archive match the title and author list of the JOSS paper. After that I think we're done 🎉
