### 2019-10-23 #1745 @ixjlyons
I've released version 0.1.1 of ndsplines and created a Zenodo entry for it: https://doi.org/10.5281/zenodo.3516465

Thank you @amritagos and @sarats for reviewing and @mbobra for editing!
