### 2019-10-24 #1745 @kthyng
Hi @sixpearls! It looks like the citations in your paper aren't working when you have more than one because you are using commas to separate multiple references — please update to use semicolons. Also, please go through your references file and use {} around any words that need to stay capitalized, like MATLAB. The {} preserve capitalization.
