### 2019-10-25 b50b2af
Fix up paper references and citations
- Specify capitalization where needed
- Use semicolons for multiple citations
