+++ paper/references.bib
@@ -1,24 +1,24 @@
 @article{margolis2017simupy,
-  title={SimuPy: A Python framework for modeling and simulating dynamical systems.},
+  title={{SimuPy}: A {Python} framework for modeling and simulating dynamical systems},
   author={Margolis, Benjamin WL},
   journal={J. Open Source Software},
   volume={2},
   number={17},
   pages={396},
   year={2017},
-  doi= {10.21105/joss.00396},
+  doi={10.21105/joss.00396},
 }
 
 @Misc{scipy,
-  author =    {Eric Jones and Travis Oliphant and Pearu Peterson and others},
-  title =     {{SciPy}: Open source scientific tools for {Python}},
-  year =      {2001--},
-  url = "http://www.scipy.org/",
-  note = {[Online; accessed 9/1/2019]}
+  title={{SciPy}: Open source scientific tools for {Python}},
+  author={Eric Jones and Travis Oliphant and Pearu Peterson and others},
+  year={2001--},
+  url="http://www.scipy.org/",
+  note={[Online; accessed 9/1/2019]}
 }
 
 @article{van2011numpy,
-  title={The NumPy array: a structure for efficient numerical computation},
+  title={The {NumPy} array: a structure for efficient numerical computation},
   author={Van Der Walt, Stefan and Colbert, S Chris and Varoquaux, Gael},
   journal={Computing in Science \& Engineering},
   volume={13},
@@ -30,7 +30,7 @@
 }
 
 @article{behnel2011cython,
-  title={Cython: The best of both worlds},
+  title={{Cython}: The best of both worlds},
   author={Behnel, Stefan and Bradshaw, Robert and Citro, Craig and Dalcin, Lisandro and Seljebotn, Dag Sverre and Smith, Kurt},
   journal={Computing in Science \& Engineering},
   volume={13},
@@ -49,7 +49,7 @@
 }
 
 @book{de1990spline,
-  title={Spline Toolbox: for Use with MATLAB: User's Guide, March 30, 1990},
+  title={Spline Toolbox: for Use with {MATLAB}: User's Guide, March 30, 1990},
   author={De Boor, Carl},
   year={1990},
   publisher={MathWorks Incorporated}
@@ -65,7 +65,7 @@
 }
 
 @inproceedings{d2019developing,
-  title={Developing an Entry Guidance and Control Design Capability Using Flaps for the Lifting Nano-ADEPT},
+  title={Developing an Entry Guidance and Control Design Capability Using Flaps for the Lifting {Nano-ADEPT}},
   author={D'Souza, Sarah N and Okolo, Wendy and Nikaido, Ben and Yount, Bryan and Tran, Jason and Margolis, Benjamin and Smith, Brandon and Cassell, Alan and Johnson, Breanna and Hibbard, Kenneth and others},
   booktitle={AIAA Aviation 2019 Forum},
   pages={2901},
@@ -81,7 +81,7 @@
 }
 
 @inproceedings{daveml,
-  title={Flight dynamic model exchange using XML},
+  title={Flight dynamic model exchange using {XML}},
   author={Jackson, E Bruce and Hildreth, Bruce},
   booktitle={AIAA Modeling and Simulation Technologies Conference and Exhibit},
   pages={4482},
