### 2019-10-25 #1745 @sixpearls
Hi @kthyng, thank you for catching that. My co-author @ixjlyons has made those corrections and it looks correct now. Let us know if there are any other corrections.
