### 2019-10-25 #1745 @kthyng
Great! Sorry I didn't realize to mention both @sixpearls and @ixjlyons.

The paper is looking good!
